# app/api/brands/route.ts
import { NextResponse } from 'next/server';
import { query } from '../../../modules/platform/db/postgres';
import { requireUser } from '../../../modules/platform/auth/current-user';
export const dynamic='force-dynamic';
export async function GET(){try{const result=await query(`select m.id,m.store_name,m.verification_status as status,count(p.id)::int as product_count from trust_merchant_profiles m left join trust_products p on p.merchant_id=m.id and p.active=true group by m.id,m.store_name,m.verification_status order by product_count desc,m.store_name asc`);return NextResponse.json({ok:true,surfaceStatus:'LIVE',source:'trust_merchant_profiles + trust_products',brands:result.rows.map(r=>({id:String(r.id),name:String(r.store_name),status:String(r.status),productCount:Number(r.product_count)}))});}catch(e){const code=e instanceof Error?e.message:'BRANDS_UNAVAILABLE';return NextResponse.json({ok:false,error:code},{status:code==='DATABASE_NOT_CONFIGURED'?503:400});}}
export async function POST(req:Request){try{const user=await requireUser(req as any);if(!['merchant','admin'].includes(String(user.role)))return NextResponse.json({ok:false,error:'MERCHANT_PERMISSION_REQUIRED'},{status:403});const b=await req.json();const targetUser=String(b.userId||user.id);if(String(user.role)==='merchant'&&targetUser!==String(user.id))return NextResponse.json({ok:false,error:'USER_OWNERSHIP_REQUIRED'},{status:403});const name=String(b.name||'').trim();const slug=(String(b.slug||name).toLowerCase().replace(/[^a-z0-9]+/g,'-').replace(/^-|-$/g,'')).slice(0,80);if(!name||!slug)throw new Error('BRAND_NAME_REQUIRED');const r=await query(`insert into trust_merchant_profiles(user_id,store_name,slug,verification_status) values($1,$2,$3,'pending') on conflict(user_id) do update set store_name=excluded.store_name,slug=excluded.slug,updated_at=now() returning id,user_id,store_name,slug,verification_status`,[targetUser,name,slug]);return NextResponse.json({ok:true,surfaceStatus:'LIVE',brand:r.rows[0]},{status:201});}catch(e){const code=e instanceof Error?e.message:'BRAND_CREATE_FAILED';return NextResponse.json({ok:false,error:code},{status:code==='DATABASE_NOT_CONFIGURED'?503:400});}}