# app/api/customer/orders/[id]/shipments/route.ts
import {NextRequest,NextResponse} from 'next/server';import {getCurrentUser} from '../../../../../../modules/platform/auth/current-user';import {listShipmentsForOrder} from '../../../../../../modules/platform/fulfillment-tracking-3';import {query} from '../../../../../../modules/platform/db/postgres';
export const dynamic='force-dynamic';export const runtime='nodejs';
export async function GET(req:NextRequest,{params}:{params:{id:string}}){const u=await getCurrentUser(req);if(!u)return NextResponse.json({ok:false,error:'AUTH_REQUIRED'},{status:401});try{const o=(await query(`select customer_id from trust_orders where id=$1`,[params.id])).rows[0];if(!o)return NextResponse.json({ok:false,error:'ORDER_NOT_FOUND'},{status:404});if(o.customer_id!==u.id&&!['admin','support','operations'].includes(u.role))return NextResponse.json({ok:false,error:'ORDER_ACCESS_DENIED'},{status:403});return NextResponse.json({ok:true,shipments:await listShipmentsForOrder(params.id)},{headers:{'Cache-Control':'no-store'}})}catch(e){return NextResponse.json({ok:false,error:e instanceof Error?e.message:'SHIPMENT_ERROR'},{status:400})}}