# app/api/returns/route.ts
import {NextRequest,NextResponse} from 'next/server';
import {getCurrentUser} from '../../../modules/platform/auth/current-user';
import {query,withPgTransaction} from '../../../modules/platform/db/postgres';
import {createReturn,listCustomerReturns,listOperationsReturns} from '../../../modules/commerce/returns/service';
export const dynamic='force-dynamic'; export const runtime='nodejs';
function privileged(role:string){return ['admin','operations','support','merchant'].includes(role);}
export async function GET(req:NextRequest){const u=await getCurrentUser(req);if(!u)return NextResponse.json({ok:false,error:'AUTH_REQUIRED'},{status:401});try{const status=req.nextUrl.searchParams.get('status')||undefined;const data=privileged(u.role)?await listOperationsReturns({query,transaction:withPgTransaction}, {status:status as any,limit:Number(req.nextUrl.searchParams.get('limit')||50)}):await listCustomerReturns({query,transaction:withPgTransaction},u.id,Number(req.nextUrl.searchParams.get('limit')||50));return NextResponse.json({ok:true,returns:data,surfaceStatus:'LIVE'},{headers:{'Cache-Control':'no-store'}});}catch(e){return NextResponse.json({ok:false,error:e instanceof Error?e.message:'RETURNS_LIST_FAILED'},{status:500});}}
export async function POST(req:NextRequest){const u=await getCurrentUser(req);if(!u)return NextResponse.json({ok:false,error:'AUTH_REQUIRED'},{status:401});const key=req.headers.get('idempotency-key')?.trim();if(!key)return NextResponse.json({ok:false,error:'IDEMPOTENCY_KEY_REQUIRED'},{status:400});try{const b=await req.json();const result=await createReturn({query,transaction:withPgTransaction} as any,{orderId:String(b.orderId||''),customerId:u.id,reasonCode:b.reasonCode,reasonNote:b.reasonNote,items:Array.isArray(b.items)?b.items:[],idempotencyKey:key});return NextResponse.json({ok:true,return:result,surfaceStatus:'LIVE'},{status:201,headers:{'Cache-Control':'no-store'}});}catch(e){const m=e instanceof Error?e.message:'RETURN_CREATE_FAILED';const status=m==='ORDER_NOT_FOUND'?404:m.includes('ACCESS_DENIED')?403:m.startsWith('INVALID_')||m.endsWith('_REQUIRED')?400:409;return NextResponse.json({ok:false,error:m},{status});}}