# docs/api/openapi-v1.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TRUST APEX OS API",
    "version": "V224.0.0",
    "description": "Generated route inventory. Request/response schemas should be extended per domain before external publication."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/api/brands": {
      "get": {
        "summary": "GET /api/brands",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Client error"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "post": {
        "summary": "POST /api/brands",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Client error"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
    "/api/customer/orders/[id]/shipments": {
      "get": {
        "summary": "GET /api/customer/orders/[id]/shipments",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Client error"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
    "/api/prevention": {
      "get": {
        "summary": "GET /api/prevention",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Client error"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "post": {
        "summary": "POST /api/prevention",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Client error"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
    "/api/returns": {
      "get": {
        "summary": "GET /api/returns",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Client error"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "post": {
        "summary": "POST /api/returns",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Client error"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
  }
}